// src/app/module--sports/models/Sports.ts
export class Sports {
    sportsId: number;
    sportsName: string;
    sportsCategory: {
      categoryId: string;
      categoryName: string;
    };
    // minAgeGroupId: {
    //   ageGroupId: number;
    //   ageGroupName: string;
    // };
    // maxAgeGroupId: {
    //   ageGroupId: number;
    //   ageGroupName: string;
    // };
    maxParticipants: number;
    images:string;
  }
  